Add LLVM patch to correct GenWidenVectorLoads function
Correct error in GenWidenVectorLoads to correctly keep track of how many bits have been loaded.

This error is e.g. visible when loading a 7-element vector with 64-bit elements on an x86-64 architecture that supports AVX instructions. The symptoms are an assertion failure in GenWidenVectorLoads (when LLVM has been built with assertions enabled).